refactor: changing dir and file name to lora
