+++ applications_user/lora/application.fam
@@ -5,7 +5,7 @@ App(
     entry_point="lora_main",
     requires=["gui"],
     stack_size=1 * 1024,
-    fap_icon="uart_demo_10px.png",
+    fap_icon="lora_10px.png",
     fap_category="GPIO",
 )
