+++ applications_user/lora/tests/test.c
@@ -2,7 +2,8 @@
 #include "unit_tests/unit_tests.h"
 
 
-int32_t run_lora_tests(void* context){
+int32_t run_lora_tests(void *context)
+{
     UNUSED(context);
     run_unit_tests();
     return 0;
